toolkit/components/extensions/test/mochitest/test_ext_contentscript.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for content script</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SpawnTask.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/ExtensionTestUtils.js"></script>
  <script type="text/javascript" src="head.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>

<script type="application/javascript;version=1.8">
"use strict";

add_task(function* test_contentscript()
{
  let extension = ExtensionTestUtils.loadExtension("content_script");
  yield extension.startup();
  info("extension loaded");

  let loadingCount = 0;
  let interactiveCount = 0;
  let completeCount = 0;
  extension.onMessage("script-run-loading", () => { loadingCount++; });
  extension.onMessage("script-run-interactive", () => { interactiveCount++; });

  let completePromise = new Promise(resolve => {
    extension.onMessage("script-run-complete", () => { completeCount++; resolve(); });
  });

  let chromeNamespacePromise = extension.awaitMessage("chrome-namespace-ok");

  let win = window.open("file_sample.html");

  yield Promise.all([waitForLoad(win), completePromise, chromeNamespacePromise]);
  info("test page loaded");

  win.close();

  is(loadingCount, 1, "document_start script ran exactly once");
  is(interactiveCount, 1, "document_end script ran exactly once");
  is(completeCount, 1, "document_idle script ran exactly once");

  yield extension.unload();
  info("extension unloaded");
});
</script>

</body>
</html>
